+++ CHANGELOG.md
@@ -22,7 +22,7 @@ All notable changes to this project will be documented in this file.
 ### :heart: Added
 * added config key backendDefaultView which will point to default backend view after login
 * added application and module version info to systemcheck page
-* added userpwd to browser
+* added userpwd and requestBody to browser
 
 ### :pencil: Changed
 * upgraded node-sass compiler and babel compiler to newest version